# rpl-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BackEnd RPL - By Dewa Permana",
        "description": "Swagger Documentation BackEnd Web RPL - Created by Dewa Permana",
        "contact": {
            "email": "dewaprmanaptr@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news/detail/{news_id}": {
            "get": {
                "tags": [
                    "Berita"
                ],
                "summary": "Get detail berita by ID",
                "operationId": "745f1fe70881b1f94de16d72f7e0656e",
                "parameters": [
                    {
                        "name": "news_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "judul": {
                                                    "type": "string",
                                                    "example": "Judul Berita"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "judul-berita_abc123"
                                                },
                                                "isi": {
                                                    "type": "string",
                                                    "example": "Konten berita lengkap..."
                                                },
                                                "gambar": {
                                                    "type": "string",
                                                    "example": "filename.jpg"
                                                },
                                                "gambar_url": {
                                                    "type": "string",
                                                    "example": "http://localhost/storage/berita/filename.jpg"
                                                },
                                                "kategori_berita": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berita tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/news": {
            "post": {
                "tags": [
                    "Berita"
                ],
                "summary": "Create new berita",
                "operationId": "4aa40d79593a38ee473e412757a002b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "kategori_id",
                                    "isi",
                                    "gambar"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "maxLength": 300,
                                        "example": "Judul Berita Baru"
                                    },
                                    "kategori_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "isi": {
                                        "type": "string",
                                        "example": "Konten berita lengkap..."
                                    },
                                    "gambar": {
                                        "description": "File gambar (max: 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berita Berhasil Diunggah!"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "judul": {
                                                    "type": "string",
                                                    "example": "Judul Berita Baru"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "judul-berita-baru_abc123"
                                                },
                                                "gambar_url": {
                                                    "type": "string",
                                                    "example": "http://localhost/storage/berita/filename.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal menambahkan berita"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari login (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}